+++ tests/testthat/setup.R
@@ -32,10 +32,10 @@ skip_if_no_py <- function() {
 }
 
 
-# helper function to skip if ergast isn't working
-skip_if_no_ergast <- function() {
-  testthat::skip_if_offline("ergast.com") # This will also skip on CRAN
-  if (is.null(get_ergast_content("current/circuits.json?limit=40"))) {
-    testthat::skip("No ergast connection available for testing")
+# helper function to skip if jolpica isn't working
+skip_if_no_jolpica <- function() {
+  testthat::skip_if_offline("jolpi.ca") # This will also skip on CRAN
+  if (is.null(get_jolpica_content("current/circuits.json?limit=40"))) {
+    testthat::skip("No Jolpica connection available for testing")
   }
 }