+++ tests/testthat/test-clear_f1_cache.R
@@ -1,5 +1,5 @@
 test_that("Cache Clearing works for memoised functions to file", {
-  skip_if_no_ergast()
+  skip_if_no_jolpica()
   # Note: cache clearing for fastf1 is not our responsibility, it's performed
   # by a call to fastf1 itself.
 
@@ -20,7 +20,7 @@ test_that("Cache Clearing works for memoised functions to file", {
 
 
 test_that("load_ciruits (off cache) works", {
-  skip_if_no_ergast()
+  skip_if_no_jolpica()
   # Set testing specific parameters - this disposes after the test finishes
   change_cache("off", persist = FALSE)
 
@@ -33,7 +33,7 @@ test_that("load_ciruits (off cache) works", {
 
 
 test_that("load_ciruits (memory cache) works", {
-  skip_if_no_ergast()
+  skip_if_no_jolpica()
 
   # Set testing specific parameters - this disposes after the test finishes
   withr::local_options("f1dataR.cache" = NULL)
@@ -55,7 +55,7 @@ test_that("load_ciruits (bad path cache) works", {
 
 
 test_that("load_ciruits (filesystem cache) works", {
-  skip_if_no_ergast()
+  skip_if_no_jolpica()
 
   # Set testing specific parameters - this disposes after the test finishes
   withr::local_options("f1dataR.cache" = NULL)